+++ src/settings.h
@@ -23,4 +23,25 @@ void     settings_load(NotesSettings *s);
 void     settings_save(const NotesSettings *s);
 char    *settings_get_config_path(void);
 
+/* SFTP/SSH connections */
+#define MAX_CONNECTIONS 32
+
+typedef struct {
+    char name[128];
+    char host[256];
+    int  port;
+    char user[128];
+    char remote_path[1024];
+    gboolean use_key;
+    char key_path[1024];
+} SftpConnection;
+
+typedef struct {
+    SftpConnection items[MAX_CONNECTIONS];
+    int count;
+} SftpConnections;
+
+void connections_load(SftpConnections *c);
+void connections_save(const SftpConnections *c);
+
 #endif